improve search box
Also included in this commit is an experimental index2 for Font Awesome
that loads in the icons from a big javascript array instead of
hard-coded into the page. The array was a conversion of the official
YAML file from FA 4.0.3 using a javascript to YAML parser
(http://nodeca.github.io/js-yaml/). Hopefully this will lead to better
search & categorization options in future, as well as a cleaner upgrade
path.